Add files via upload
The output file is a temporary file that is used and the schedule file has to be cleared before printing out the default schedule. Changes will be saved if the file is not empty but to restart the schedule, clear the information in the file, and print a new schedule to become the default.